+++ trunk/GoJava7/dobritsa/KickstarterD_Web_DB/src/main/java/ua/com/goit/gojava7/kickstarter/servlet/QuestionServlet.java
@@ -37,7 +37,7 @@ public class QuestionServlet extends HttpServlet {
 			throws ServletException, IOException {
 		
 		log.info("doPost()...");		
-		int projectId = Integer.parseInt(request.getParameter("projectId"));
+		Long projectId = Long.parseLong(request.getParameter("projectId"));
 		String textQuestin = request.getParameter("question");
 
 		if (!request.getParameter("question").isEmpty()) {
